+++ axlearn/cloud/gcp/jobs/launch.py
@@ -7,30 +7,59 @@ The launch command provides a few benefits:
 2. Provides utilities for registering custom launcher implementations.
 
 The launch flow depends on the launcher being used. Each launcher must define a "matcher" function
-that decides, for a given CLI action (e.g. 'start') and instance type (e.g. 'tpu-v4-8'), whether the
-launcher can be used. See `_LAUNCHERS` for a full list, and `BastionManagedGKEJob` for an example.
+that decides, for a given CLI action (e.g. 'start'), whether the launcher can be used. Run with
+`--help` to display the list of launchers.
 
-Possible actions: [start|update|stop|list]
+Further, the launch command supports specifying the `runner` implementation to use for deploying and
+monitoring the job on the bastion. The API follows a similar API as training, where
+`--runner_module` points to a module exposing `named_runner_configs`, and `--runner_name` specifies
+a specific runner to be used. Run with `--help` to display a list of runners, or with
+`--runner_name` and `--help` to see all possible flags for a specific runner.
+
+Some additional background on runners can be found in the infrastructure documentation:
+https://github.com/apple/axlearn/blob/main/docs/04-infrastructure.md#bastion-runners
+
+Possible actions: [start|update|stop|list|run]
 
     Start: submits a job to the queue.
     Update: updates a job without resubmission.
     Stop: stops the job or removes a job from the queue.
     List: lists jobs and their statuses.
+    Run: run the job locally for debugging.
 
-Additional flags may be supplied based on --instance_type. Run with --help to see all flags.
-Commands are parsed from the positional arguments (so anything after a trailing `--`).
+Additional flags may be supplied depending on the launcher. Run with --help to see all flags.
+Commands are either parsed from the positional arguments (so anything after a trailing `--`), or
+from flags, depending on the runner being used.
 
 Examples:
 
+    # View possible runners.
+    axlearn gcp launch --help
+
     # View all launch flags for the given instance type.
+    # This internally attempts to infer the --runner_name automatically.
     axlearn gcp launch --instance_type=tpu-v4-8 --help
 
-    # Simple TPU launch. Internally submits `axlearn gcp tpu ...` to bastion to run, instead of
-    # running locally.
-    axlearn gcp launch --instance_type=tpu-v4-32 -- python3 my_script.py
+    # View all launch flags for a specific runner.
+    axlearn gcp launch --runner_name=gke_tpu_single --help
 
-    # Launch with extra dependencies.
-    axlearn gcp launch --instance_type=tpu-v4-32 --bundler_spec=extras=dev -- python3 my_script.py
+    # Simple TPU launch, running locally.
+    axlearn gcp launch run --instance_type=tpu-v4-8 \
+        --bundler_spec=image=tpu --bundler_spec=dockerfile=Dockerfile -- python3 my_script.py
+
+    # Simple TPU launch, submitting to bastion to run instead of locally.
+    axlearn gcp launch --instance_type=tpu-v4-32 \
+        --bundler_spec=image=tpu --bundler_spec=dockerfile=Dockerfile -- python3 my_script.py
+
+    # Simple TPU launch, targeting a specific runner.
+    # By default, this users `axlearn.cloud.gcp.runners` as the runner module.
+    axlearn gcp launch --runner_name=gke_tpu_single --instance_type=tpu-v4-32 \
+        --bundler_spec=image=tpu --bundler_spec=dockerfile=Dockerfile -- python3 my_script.py
+
+    # Simple TPU launch, targeting a specific runner in a custom module.
+    # In this case, the available flags depend on the custom runner implementation.
+    axlearn gcp launch --runner_module=my.custom.module --runner_name=my_custom_runner \
+        --instance_type=my-custom-instance ... -- python3 my_script.py
 
     # Dry-run: prints the job config without launching.
     axlearn gcp launch --instance_type=tpu-v4-32 --dry_run -- python3 my_script.py
@@ -45,8 +74,12 @@ Examples:
     # Update an existing job without resubmission.
     axlearn gcp launch update --instance_type=tpu-v4-32 ... -- python3 my_script2.py
 
+    # To list running jobs.
+    axlearn gcp launch list
+
     # To stop a job.
-    axlearn gcp launch stop --name=... --instance_type=tpu
+    axlearn gcp launch stop --name=...
+
 
 More on the Update command:
 
@@ -66,6 +99,7 @@ More on the Update command:
 """
 # pylint: disable=redefined-outer-name,protected-access
 
+import enum
 import functools
 import os
 import shlex
@@ -73,42 +107,50 @@ import sys
 import tempfile
 from collections.abc import Sequence
 from datetime import datetime, timezone
-from typing import Callable, NamedTuple, Optional, Protocol, TextIO
+from importlib import import_module
+from typing import Callable, NamedTuple, Optional, TextIO
 
 from absl import app, flags, logging
 
 from axlearn.cloud.common.bastion import BastionDirectory
 from axlearn.cloud.common.bastion import Job as BastionJob
 from axlearn.cloud.common.bastion import new_jobspec, serialize_jobspec
+from axlearn.cloud.common.bundler import Bundler, bundler_flags, get_bundler_config
 from axlearn.cloud.common.quota import QUOTA_CONFIG_PATH, get_user_projects
 from axlearn.cloud.common.scheduler import JobMetadata
 from axlearn.cloud.common.types import JobSpec, ResourceMap
 from axlearn.cloud.common.utils import (
+    FlagConfigurable,
     configure_logging,
+    define_flags,
+    from_flags,
     generate_job_id,
     generate_job_name,
     infer_cli_name,
+    infer_resources,
     parse_action,
 )
+from axlearn.cloud.gcp import runners
+from axlearn.cloud.gcp.bundler import CloudBuildBundler
 from axlearn.cloud.gcp.config import default_project, default_zone, gcp_settings
-from axlearn.cloud.gcp.job import Job
-from axlearn.cloud.gcp.jobs import gke_runner
 from axlearn.cloud.gcp.jobs.bastion_vm import bastion_root_dir, shared_bastion_name
-from axlearn.cloud.gcp.jobs.gke_runner import JobType, job_type_flags
 from axlearn.cloud.gcp.jobs.launch_utils import (
     JobsToTableFn,
+    Matcher,
+    infer_gcp_api,
+    infer_module_qualname,
     jobs_table,
-    match_by_regex,
+    match_gcp_api,
     project_usage_table,
-    serialized_flags_for_job,
     user_usage_table,
     validate_resource_flags,
     with_k8s_jobset_state,
 )
-from axlearn.cloud.gcp.tpu import infer_tpu_resources, infer_tpu_type, infer_tpu_workers
-from axlearn.cloud.gcp.utils import GCPAPI, catch_auth, load_kube_config
+from axlearn.cloud.gcp.runners.base import BaseRunnerJob
+from axlearn.cloud.gcp.utils import GCPAPI, catch_auth, load_kube_config, running_from_vm
 from axlearn.common.config import (
     REQUIRED,
+    ConfigBase,
     ConfigOr,
     Required,
     config_class,
@@ -117,55 +159,44 @@ from axlearn.common.config import (
 )
 
 FLAGS = flags.FLAGS
+_ACTIONS = ["start", "stop", "update", "list", "run"]
+_RUNNER_ACTIONS = ["start", "update", "run"]  # Actions which require a runner.
 
 
-class _Matcher(Protocol):
-    def __call__(self, *, action: str, instance_type: str, gcp_api: str, job_type: str) -> bool:
-        pass
-
-
-class Launcher(NamedTuple):
-    """A job launcher.
+class _RegistryMember(NamedTuple):
+    """A member of a class registry.
 
     Consists of:
-    * job_cls:
-        A Job class which defines, at minimum, `_execute()`, `_list()` and `_delete()` methods,
-        invoked for "start", "list", "stop" actions respectively. It can optionally define a
-        `define_flags()` classmethod, which can be used for registering launch flags. These flags
-        will automatically be printed when invoking with --help, and will automatically be provided
-        to `Job.from_flags` when instantiating the Job.
-    * matcher:
-        A `_Matcher` (or a config instantiating thereof), used to decide whether the launcher is
-        applicable for a given action and instance type.
-    * description:
-        Human-readable description of the launcher. Printed e.g. if no launchers are matched.
+    * cfg: A `FlagConfigurable` config.
+    * matcher: A `_Matcher` (or a config instantiating thereof), used to decide whether the class
+        is applicable for a given action and flag values.
+    * description: Human-readable description of the job. Printed e.g. if no classes are matched.
     """
 
-    # We take a class here instead of a config, since a materialized config is often not available
-    # at the time of registry. Instead, we often construct the config via `job_cls.from_flags`.
-    job_cls: type[Job]
+    # A partially defined config.
+    config: FlagConfigurable.Config
     # A config is usually more print-friendly, but not strictly required.
-    matcher: ConfigOr[_Matcher]
+    matcher: ConfigOr[Matcher]
     description: str
 
 
-class BaseBastionManagedJob(Job):
+# A sequence of registry members, to be matched in the given order. The first match is used.
+_Registry = Sequence[_RegistryMember]
+
+
+# TODO(markblee): Remove GCP-specific logic and move to a common launch file.
+class BaseBastionManagedJob(FlagConfigurable):
     """A base job definition for jobs managed by a bastion.
 
     It provides functionality to submit, delete, and list bastion jobs, but is agnostic to specific
-    resource types. At minimum, subclasses should override `runner` and `_resources()` to specify
-    the implementation of the job executed by the bastion, as well as resources used by the job,
-    which will be used for quota management and scheduling.
+    resource types. At minimum, subclasses should override `runner` configure `resources` used by
+    the job, which will be used for quota management and scheduling.
 
     See `BastionManagedGKEJob` as an example.
     """
 
-    # Runner class, a subclass of Job that runs locally on the bastion.
-    # Used to infer launch flags and launch command.
-    runner: type[Job]
-
     @config_class
-    class Config(Job.Config):
+    class Config(FlagConfigurable.Config):
         """Configures BaseBastionManagedJob."""
 
         # Name of the job. It's used for bastion job management as well as the name of the runner.
@@ -176,9 +207,7 @@ class BaseBastionManagedJob(Job):
         env_id: Optional[str] = None
         # Where to run the remote job.
         zone: Required[str] = REQUIRED
-        # Instance type to launch.
-        instance_type: Required[str] = REQUIRED
-        # Bastion name.
+        # Bastion name. TODO(markblee): Move to bastion_dir and infer the root dir.
         bastion_name: Required[str] = REQUIRED
         # Bastion dir.
         bastion_dir: BastionDirectory.Config = BastionDirectory.default_config()
@@ -190,8 +219,10 @@ class BaseBastionManagedJob(Job):
         priority: Required[int] = REQUIRED
         # Output directory for job logs.
         output_dir: Required[str] = REQUIRED
+        # Bundler config.
+        bundler: Optional[Bundler.Config] = None
         # Runner executed by the bastion.
-        runner: Required[Job.Config] = REQUIRED
+        runner: Optional[BaseRunnerJob.Config] = None
         # One or more functions for displaying `list` information. Each is invoked with the bastion
         # jobs and is expected to return a printable table.
         output_tables: Sequence[ConfigOr[JobsToTableFn]] = [
@@ -200,23 +231,14 @@ class BaseBastionManagedJob(Job):
             project_usage_table,
         ]
         # Resources used by the job.
-        resources: ConfigOr[ResourceMap[int]] = {}
-
-    @classmethod
-    def with_runner(cls, runner: type[Job]):
-        return type(f"{cls.__name__}_{runner.__name__}", (cls,), {"runner": runner})
-
-    @classmethod
-    def validate_runner(cls):
-        if cls.runner is None:
-            raise ValueError(f"A BastionManagedJob should subclass {cls} and define `runner`.")
+        resources: Callable[[ConfigBase], ResourceMap[int]] = infer_resources
 
     @classmethod
     def define_flags(cls, fv: flags.FlagValues):
         """Defines launch flags on the provided flag_values."""
-        cls.validate_runner()
         super().define_flags(fv)
         common_kwargs = dict(flag_values=fv, allow_override=True)
+        bundler_flags(required=False, **common_kwargs)
         flags.DEFINE_string("name", None, "Name of the job.", **common_kwargs)
         flags.DEFINE_string("bastion", None, "Name of bastion VM to use.", **common_kwargs)
         flags.DEFINE_integer(
@@ -244,7 +266,6 @@ class BaseBastionManagedJob(Job):
             "If specified, the directory to store outputs (such as logs).",
             **common_kwargs,
         )
-        cls.runner.define_flags(fv)
 
     @classmethod
     def set_defaults(cls, fv):
@@ -253,56 +274,66 @@ class BaseBastionManagedJob(Job):
         fv.set_default("name", fv.name or generate_job_name())
 
     @classmethod
-    def from_flags(cls, fv: flags.FlagValues, *, command: str, action: str, **kwargs) -> Config:
+    def from_flags(
+        cls, fv: flags.FlagValues, *, action: str, command: Optional[str] = None, **kwargs
+    ) -> Config:
         """Constructs config from flags defined by `define_flags()`.
 
         Args:
             fv: Flag values (e.g., FLAGS).
-            command: The user-supplied command, i.e. everything after `--` as a string.
             action: Action requested by the user.
+            command: The user-supplied command, i.e. everything after `--` as a string.
+                If None, command should be specified via flags (e.g., in the case of multiple
+                subcommands).
             kwargs: Optional key/values to set on the config.
 
         Returns:
             The job config.
         """
-        cls.validate_runner()
         cfg: BaseBastionManagedJob.Config = super().from_flags(fv, **kwargs)
         if not cfg.bastion_name:
-            cfg.bastion_name = fv.bastion or shared_bastion_name(fv, gcp_api=_gcp_api(fv))
+            cfg.bastion_name = fv.bastion or shared_bastion_name(fv, gcp_api=infer_gcp_api(fv))
         cfg.bastion_dir.root_dir = bastion_root_dir(cfg.bastion_name, fv=fv)
         # Default output_dir depends on the final value of --name.
         if not cfg.output_dir:
             cfg.output_dir = f"gs://{gcp_settings('ttl_bucket', fv=fv)}/axlearn/jobs/{fv.name}"
         # Construct runner only for start and update.
-        if action in ("start", "update"):
-            cfg.runner = cls.runner.from_flags(fv, command=command)
-            runner_flags = " ".join(serialized_flags_for_job(fv, cls.runner))
-            cfg.command = (
-                f"python3 -m {cls.runner.__module__} {action} {runner_flags} "
-                f"--job_type={fv.job_type} "
-                f"-- {command}"
+        if action in _RUNNER_ACTIONS:
+            # We construct a bundler and propagate to runner during instantiate, ensuring the
+            # bundling is consistent between local and bastion.
+            cfg.bundler = get_bundler_config(
+                bundler_type=fv.bundler_type, spec=fv.bundler_spec, fv=fv
             )
+            # Build launch command. We take the same flags provided to this module and run the
+            # command again, this time on the bastion with action="run".
+            # For backwards compatibility with legacy behavior where command is specified with argv
+            # after `--`, we take any command supplied via `kwargs` and use argv.
+            # TODO(markblee): Handle quoting for flag commands that contain spaces.
+            own_flags = " ".join(fv.flags_into_string().split("\n"))
+            cfg.command = f"python3 -m {infer_module_qualname(cls)} run {own_flags}"
+            if command is not None:
+                cfg.command = f"{cfg.command} -- {command}"
         else:
-            cfg.runner = None
             cfg.command = None
         return cfg
 
     def __init__(self, cfg: Config):
         super().__init__(cfg)
         cfg = self.config
-        if not (cfg.instance_type and cfg.output_dir):
-            raise ValueError("instance_type, output_dir cannot be empty")
+        if not cfg.output_dir:
+            raise ValueError("output_dir cannot be empty")
         self._bastion_dir: BastionDirectory = cfg.bastion_dir.instantiate()
-        self._runner = None
+        self._bundler = self._runner = None
         if cfg.command is not None:
-            self._runner: Job = cfg.runner.instantiate()
+            self._bundler: Bundler = cfg.bundler.instantiate()
+            self._runner: BaseRunnerJob = cfg.runner.instantiate(bundler=self._bundler)
         self._output_tables = maybe_instantiate(cfg.output_tables)
 
-    def _delete(self):
-        """Submits a delete request to bastion."""
+    def cancel(self):
+        """Submits a cancel request to bastion."""
         self._bastion_dir.cancel_job(self.config.name)
 
-    def _list(self, output_file: Optional[TextIO] = None) -> dict[str, BastionJob]:
+    def list(self, output_file: Optional[TextIO] = None) -> dict[str, BastionJob]:
         """Lists running jobs and optionally prints them in tabular format.
 
         Args:
@@ -317,13 +348,23 @@ class BaseBastionManagedJob(Job):
             print(fn(jobs), file=output_file)
         return jobs
 
-    def _execute(self) -> JobSpec:
-        """Submits the command to bastion."""
+    def run(self):
+        """Runs the command locally.
+
+        Note that 'locally' can also mean locally on the bastion: `submit()` takes the runner
+        command, ships it to the bastion, and the bastion executes it with `run()`.
+        """
         cfg: BaseBastionManagedJob.Config = self.config
+        from_vm = running_from_vm()
+
+        if not from_vm:
+            os.environ["BASTION_TIER"] = "disabled"
+            self._bundler.bundle(cfg.name)
 
         # Check for group membership if --project_id is provided.
-        # TODO(markblee): Make this check at the bastion level.
         if cfg.project_id:
+            if not from_vm:
+                logging.warning("Supplying --project_id has no purpose if not running in bastion.")
             quota_file = (
                 f"gs://{gcp_settings('private_bucket')}/{cfg.bastion_name}/{QUOTA_CONFIG_PATH}"
             )
@@ -333,12 +374,12 @@ class BaseBastionManagedJob(Job):
                     f"User '{cfg.user_id}' is not a member of the project '{cfg.project_id}'. "
                     f"Instead, user '{cfg.user_id}' is a member of: {user_projects}"
                 )
+        self._runner.execute()
 
-        # TODO(markblee): Simplify when merging launch command.
-        # pytype: disable=attribute-error
-        if self._runner and self._runner._bundler:
-            self._runner._bundler.bundle(cfg.name)
-        # pytype: enable=attribute-error
+    def submit(self) -> JobSpec:
+        """Submits the command to bastion."""
+        cfg: BaseBastionManagedJob.Config = self.config
+        self._bundler.bundle(cfg.name)
 
         logging.info("Starting run for job name %s", cfg.name)
         logging.info("Command: %s", cfg.command)
@@ -348,7 +389,7 @@ class BaseBastionManagedJob(Job):
                 user_id=cfg.user_id,
                 project_id=cfg.project_id or "none",
                 creation_time=datetime.now(timezone.utc),
-                resources=maybe_instantiate(cfg.resources),
+                resources=cfg.resources(cfg),
                 priority=cfg.priority,
                 job_id=job_id,
             )
@@ -360,20 +401,17 @@ class BaseBastionManagedJob(Job):
             f"gsutil cat {os.path.join(self._bastion_dir.logs_dir, cfg.name)}\n"
             f"\nStop/cancel the job with:\n"
             f"{infer_cli_name()} gcp launch stop "
-            f"--name={cfg.name} --bastion={cfg.bastion_name} --instance_type={cfg.instance_type} "
-            f"--env_id={cfg.env_id}\n"
+            f"--name={cfg.name} --bastion={cfg.bastion_name} --env_id={cfg.env_id}\n"
             "\nCheck job history with:\n"
             f"{infer_cli_name()} gcp bastion history "
-            f"--name={cfg.bastion_name} --env_id={cfg.env_id} "
-            f"--job_name={cfg.name}"
+            f"--name={cfg.bastion_name} --env_id={cfg.env_id} --job_name={cfg.name}\n"
             "\nCheck project history with:\n"
             f"{infer_cli_name()} gcp bastion history "
-            f"--name={cfg.bastion_name} --env_id={cfg.env_id} "
-            f"{cfg.project_id or ''}"
+            f"--name={cfg.bastion_name} --env_id={cfg.env_id} {cfg.project_id or ''}"
         )
         return jobspec
 
-    def _update(self) -> JobSpec:
+    def update(self) -> JobSpec:
         """Update an existing job without resubmission.
 
         This will fetch the existing job from Bastion, change
@@ -386,12 +424,8 @@ class BaseBastionManagedJob(Job):
 
         # Get current job spec.
         job_spec = self._bastion_dir.get_job(job_name=cfg.name)
-
-        # TODO(markblee): Simplify when merging launch command.
-        # pytype: disable=attribute-error
-        if self._runner and self._runner._bundler:
-            self._runner._bundler.bundle(cfg.name)
-        # pytype: enable=attribute-error
+        # Update the bundle.
+        self._bundler.bundle(cfg.name)
 
         logging.info("Starting update for job name %s", cfg.name)
         logging.info("Command: %s", cfg.command)
@@ -407,9 +441,7 @@ class BaseBastionManagedJob(Job):
 
         logging.info("Updated jobspec: %s", job_spec)
 
-        jobspec = self._bastion_dir.update_job(cfg.name, job_spec=job_spec)
-
-        return jobspec
+        return self._bastion_dir.update_job(cfg.name, job_spec=job_spec)
 
 
 class BastionManagedGKEJob(BaseBastionManagedJob):
@@ -424,18 +456,16 @@ class BastionManagedGKEJob(BaseBastionManagedJob):
             project: Used for load_kube_config.
             zone: Used to infer total quota.
             cluster: K8s cluster.
-            num_replicas: Number of replicas.
         """
 
         namespace: str = "default"
         project: Required[str] = REQUIRED
         zone: Required[str] = REQUIRED
         cluster: Required[str] = REQUIRED
-        num_replicas: int = 1
 
     @classmethod
     def define_flags(cls, fv: flags.FlagValues):
-        """Defines launch flags using tpu_runner."""
+        """Defines launch flags."""
         super().define_flags(fv)
         common_kwargs = dict(flag_values=fv, allow_override=True)
         flags.DEFINE_string("project", None, "The GCP project name.", **common_kwargs)
@@ -446,29 +476,22 @@ class BastionManagedGKEJob(BaseBastionManagedJob):
     @classmethod
     def set_defaults(cls, fv: flags.FlagValues):
         super().set_defaults(fv)
-        # Don't override `name` if already specified, since the default is non-deterministic.
-        fv.set_default("name", fv.name or generate_job_name())
         fv.set_default("project", default_project())
         fv.set_default("zone", default_zone())
         fv.set_default("namespace", "default")
+        # While bundler is only instantiated for certain actions, it doesn't hurt to specify a
+        # default even if it's not used.
+        fv.set_default("bundler_type", CloudBuildBundler.TYPE)
 
     @classmethod
-    def from_flags(cls, fv: flags.FlagValues, *, command: str, action: str, **kwargs) -> Config:
-        cfg: BastionManagedGKEJob.Config = super().from_flags(
-            fv, command=command, action=action, **kwargs
-        )
+    def from_flags(cls, fv: flags.FlagValues, *, action: str, **kwargs) -> Config:
+        cfg: BastionManagedGKEJob.Config = super().from_flags(fv, action=action, **kwargs)
         cfg.cluster = cfg.cluster or gcp_settings("gke_cluster", required=False, fv=fv)
         cfg.output_tables = [
             with_k8s_jobset_state(jobs_table, namespace=cfg.namespace),
             user_usage_table,
             project_usage_table,
         ]
-        # Use config_for_function to delay instantiation unless needed, e.g., it may not be
-        # necessary for list action.
-        cfg.resources = config_for_function(infer_tpu_resources).set(
-            instance_type=fv.instance_type,
-            num_replicas=fv.num_replicas,
-        )
         return cfg
 
     def __init__(self, cfg: Config):
@@ -476,22 +499,12 @@ class BastionManagedGKEJob(BaseBastionManagedJob):
         # Ensure that the user has installed GKE plugin.
         load_kube_config(project=cfg.project, zone=cfg.zone, cluster=cfg.cluster)
 
-    def _execute(self) -> JobSpec:
+    def submit(self) -> JobSpec:
         """Submits the command to bastion."""
         cfg: BastionManagedGKEJob.Config = self.config
-        try:
-            num_workers = infer_tpu_workers(infer_tpu_type(cfg.instance_type))
-            worker_log = f"{infer_cli_name()} gcp logs --name={cfg.name} --worker=0"
-            print(
-                f"\nOnce started, view TPU log outputs with:\n{worker_log}\n"
-                "Replace `--worker=0` with `--worker={idx}` "
-                f"where idx is between [0, {num_workers}).\n"
-            )
-        except ValueError:
-            logging.warning(
-                "Failed to infer number of workers for instance_type: %s.", cfg.instance_type
-            )
-        job_spec = super()._execute()
+        worker_log = f"{infer_cli_name()} gcp logs --name={cfg.name} --worker=0"
+        print(f"\nOnce started, view log outputs with:\n{worker_log}\n")
+        job_spec = super().submit()
         print(
             "\nView running pods with:\nkubectl get pods\n"
             "\nNote that the job may take a few minutes to start."
@@ -500,117 +513,135 @@ class BastionManagedGKEJob(BaseBastionManagedJob):
 
 
 # Launchers specified here will be tried (in the given order) when launching a given instance type.
-_LAUNCHERS = [
-    Launcher(
-        job_cls=BastionManagedGKEJob.with_runner(gke_runner.FlinkGKERunnerJob),
-        matcher=config_for_function(match_by_regex).set(
-            match_regex=None,
-            gcp_api=None,
-            job_type=JobType.FLINK.value,
-        ),
-        description="Supports running Flink jobs on GKE TPU",
-    ),
-    # TPU GKE launcher.
-    Launcher(
-        job_cls=BastionManagedGKEJob.with_runner(gke_runner.TPUGKERunnerJob),
-        matcher=config_for_function(match_by_regex).set(
-            match_regex=dict(start=r"tpu-v.+-(\d)+", update=r"tpu.*", list=r"tpu.*", stop=r"tpu.*"),
-            gcp_api=GCPAPI.GKE.value,
-            job_type=JobType.DEFAULT.value,
-        ),
-        description=(
-            "Supports launching TPU jobs via GKE. "
-            "For 'start' or 'update', provide the instance type, e.g. --instance_type=tpu-v4-8. "
-            "For 'list' or 'stop', provide the accelerator type, e.g. --instance_type=tpu."
-        ),
-    ),
+_LAUNCHERS: _Registry = [
+    # GKE launcher.
+    _RegistryMember(
+        config=BastionManagedGKEJob.default_config(),
+        matcher=config_for_function(match_gcp_api).set(gcp_api=GCPAPI.GKE.value),
+        description="Supports launching jobs to GKE.",
+    )
 ]
 
 
+def _registry_help_string(registry: Sequence[_RegistryMember]) -> str:
+    return "\n".join(
+        [
+            f"Job: {member.config.klass.__name__}\n"
+            f"Description: {member.description}\n"
+            f"Matcher: {member.matcher}\n"
+            for member in registry
+        ]
+    )
+
+
+def _get_runner_or_exit(fv: flags.FlagValues = FLAGS) -> BaseRunnerJob.Config:
+    """Infers the runner config from runner_module and runner_name."""
+    module = import_module(fv.runner_module)
+    return module.named_runner_configs(getattr(fv, "runner_name", None))
+
+
 def _get_launcher_or_exit(
-    *, action: str, instance_type: str, gcp_api: str, flag_values: flags.FlagValues = FLAGS
-) -> Launcher:
-    """Retrieves launcher by matching instance_type and gcp_api.
+    *, action: str, require_runner: bool = True, flag_values: flags.FlagValues = FLAGS
+) -> BaseBastionManagedJob.Config:
+    """Retrieves launcher by matching `action` and flags.
 
     If there are multiple matches, the first one in the registry is returned.
     """
-    # Identify launcher from instance type.
     for launcher in _LAUNCHERS:
         m = maybe_instantiate(launcher.matcher)
-        if m(
-            action=action,
-            instance_type=instance_type,
-            gcp_api=gcp_api,
-            job_type=flag_values.job_type,
-        ):
-            return launcher
+        if m(action=action, flag_values=flag_values):
+            break
+    else:
+        raise app.UsageError(
+            f"Failed to find a match in the registry for {action=}. "
+            f"The provided flags are:\n\n{flag_values.flag_values_dict()}.\n\n"
+            f"The registered options are:\n\n{_registry_help_string(_LAUNCHERS)}"
+        )
+    # Make a copy of the config to avoid modifying global launcher.
+    launch_cfg: BaseBastionManagedJob.Config = launcher.config.clone()
+    try:
+        flag_values.runner_name = _infer_runner_name(flag_values)
+        launch_cfg.runner = _get_runner_or_exit(flag_values)
+    except Exception:  # pylint: disable=broad-except
+        if require_runner:
+            raise
+        logging.warning(
+            "Failed to infer runner name. Proceeding since require_runner=%s", require_runner
+        )
+    return launch_cfg
 
-    launchers = "\n".join(
-        [
-            f"Job: {launcher.job_cls.__name__}\n"
-            f"Description: {launcher.description}\n"
-            f"Matcher: {launcher.matcher}\n"
-            for launcher in _LAUNCHERS
-        ]
-    )
-    raise app.UsageError(
-        f"Don't know how to launch {instance_type=} for {action=} and {gcp_api=}.\n"
-        f"The registered launchers are:\n\n{launchers}"
-    )
 
-
-def _gcp_api(fv: flags.FlagValues = FLAGS) -> str:
-    if getattr(fv, "gcp_api", None) is not None:
-        return fv.gcp_api.lower()
-    # The return value depends on --zone, so cannot be set as the default value of fv.gcp_api.
-    return gcp_settings(
-        "launch_gcp_api", default=GCPAPI.GKE.lower(), required=False, fv=fv
-    )  # pytype: disable=bad-return-type
+def _parse_command_from_argv(argv: Sequence[str], *, action: str) -> str:
+    # Parse the command from argv. Note that argv may or may not contain action, so we explicitly
+    # look for '--' and extract all args after it.
+    command = ""
+    for i, arg in enumerate(argv):
+        if arg.strip() == "--":
+            # For "run", we don't do additional quoting as the joined command is supplied directly
+            # to the runner (e.g. k8s manifest), without going through the shell.
+            # For "submit", we quote since the command will be shipped to the bastion, at which
+            # point it goes through the shell before "run".
+            # TODO(markblee): Improve/simplify this behavior.
+            command = " ".join(argv[i + 1 :]) if action == "run" else shlex.join(argv[i + 1 :])
+            break
+    return command
 
 
 @catch_auth
-def main(_):
-    if FLAGS.instance_type is None:
-        raise app.UsageError("--instance_type is required.")
-
-    action = parse_action(sys.argv, options=["start", "stop", "update", "list"], default="start")
-    launcher = _get_launcher_or_exit(
-        action=action,
-        instance_type=FLAGS.instance_type,
-        gcp_api=_gcp_api(),
-        flag_values=FLAGS,
+def main(_, fv: flags.FlagValues = FLAGS):
+    action = parse_action(sys.argv, options=_ACTIONS, default="start")
+    cfg = _get_launcher_or_exit(
+        action=action, require_runner=action in _RUNNER_ACTIONS, flag_values=fv
     )
 
-    # Parse the command from argv. Note that argv may or may not contain action, so we explicitly
-    # look for '--' and extract all args after it. Use sys.argv instead of argv from params, since
-    # the param argv has '--' stripped.
-    command = ""
-    for i, arg in enumerate(sys.argv):
-        if arg.strip() == "--":
-            command = shlex.join(sys.argv[i + 1 :])
-            break
-    cfg = launcher.job_cls.from_flags(FLAGS, command=command, action=action)
-    logging.info("Launcher config:\n%s", cfg)
+    # Use sys.argv instead of argv from params, since the param argv has '--' stripped.
+    command = _parse_command_from_argv(sys.argv, action=action)
+
+    # If command is supplied, treat as a single --command flag.
+    # In all other cases, we expect one or more commands specified as flags.
+    from_flags_kwargs = dict(action=action)
+    if command:
+        from_flags_kwargs["command"] = command
+
+    cfg = from_flags(cfg, fv, **from_flags_kwargs)
+    logging.info("Action: %s, Launcher config:\n%s", action, cfg)
     job: BaseBastionManagedJob = cfg.instantiate()
-    if FLAGS.dry_run:
-        print(f"Action: {action}\nJob config:\n{job.config}")
+    if fv.dry_run:
         return
 
     if action == "start":
-        job._execute()
+        job.submit()
     elif action == "list":
-        job._list()
+        job.list()
     elif action == "stop":
-        job._delete()
+        job.cancel()
     elif action == "update":
-        job._update()
+        job.update()
+    elif action == "run":
+        try:
+            job.run()
+        except KeyboardInterrupt:
+            # This is mostly for local debugging, if executing the runner locally.
+            logging.info("Got keyboard interrupt, deleting resources...")
+            job._runner._delete()  # pytype: disable=attribute-error
     else:
         raise app.UsageError(f"Unsupported action {action}")
 
 
+class _JobType(enum.Enum):
+    """Represents possible values for `--job_type`.
+
+    This is used for selecting the type of runner to use, in the cases where the same instance
+    type can map to multiple possible runners.
+    """
+
+    # NOTE: do not add more types here. This will be removed in the future.
+    DEFAULT = "default"
+    FLINK = "flink"
+
+
 def _prelaunch_flags(fv: flags.FlagValues = FLAGS):
     """Flags necessary for `_get_launcher_or_exit`."""
-    flags.DEFINE_string("instance_type", None, "Instance type to launch.", flag_values=fv)
     # pytype: disable=missing-parameter
     flags.DEFINE_enum(
         "gcp_api",
@@ -623,35 +654,84 @@ def _prelaunch_flags(fv: flags.FlagValues = FLAGS):
     flags.DEFINE_bool(
         "dry_run", False, "Output job config and exit without running.", flag_values=fv
     )
-    job_type_flags(fv)
+    flags.DEFINE_string("runner_module", runners.__name__, "The runner module.", flag_values=fv)
+    flags.DEFINE_string("runner_name", None, "The runner module.", flag_values=fv)
+
+    # TODO(markblee): These flags are for backwards compatibility.
+    flags.DEFINE_string(
+        "instance_type",
+        None,
+        "(Deprecated) Instance type to launch. "
+        "Please use --runner_module and --runner_name instead.",
+        flag_values=fv,
+    )
+    flags.DEFINE_enum(
+        "job_type",
+        _JobType.DEFAULT.value,
+        [v.value for v in _JobType],
+        "(Deprecated) Job type. Please use --runner_module and --runner_name instead.",
+        flag_values=fv,
+    )
 
 
-def _private_flags():
+def _infer_runner_name(fv: flags.FlagValues = FLAGS) -> str:
+    """Infers the --runner_name from legacy flags --instance_type and --job_type."""
+    if fv.runner_name:
+        return fv.runner_name
+
+    logging.warning("--runner_name was not provided. Will attempt to infer automatically.")
+    runner_name = None
+    if getattr(fv, "job_type", None) == _JobType.FLINK.value:
+        runner_name = "gke_tpu_flink"
+    elif getattr(fv, "instance_type", None):
+        if fv.instance_type.startswith("tpu"):
+            runner_name = "gke_tpu_single"
+        elif fv.instance_type.startswith("gpu-a3"):
+            runner_name = "gke_gpu_a3_single"
+        else:
+            raise app.UsageError(
+                f"Unable to infer --runner_name from --instance_type={fv.instance_type}; "
+                "Please specify --runner_name explicitly, or use a different --instance_type."
+            )
+    else:
+        raise app.UsageError(
+            "Unable to infer --runner_name. "
+            "Please specify --runner_name explicitly, or specify a valid --instance_type."
+        )
+
+    assert runner_name is not None
+    logging.info("Inferred runner name: %s", runner_name)
+    return runner_name
+
+
+def _private_flags(
+    fv: flags.FlagValues = FLAGS,
+    prelaunch_fn: Callable[[flags.FlagValues], None] = _prelaunch_flags,
+):
     """Defines all launch flags, and amends `app.usage` with additional launch help info."""
 
-    _prelaunch_flags()
-    FLAGS(sys.argv, known_only=True)
+    prelaunch_fn(fv)
+    fv(sys.argv, known_only=True)
 
-    launch_help = None
-    # Allow instance_type to be None when running --help without any flags. On the other hand, if
-    # instance_type is provided when running --help, we show additional help info.
-    if FLAGS.instance_type:
-        action = parse_action(
-            sys.argv, options=["start", "update", "stop", "list"], default="start"
-        )
-        launcher = _get_launcher_or_exit(
-            action=action,
-            instance_type=FLAGS.instance_type,
-            gcp_api=_gcp_api(),
-            flag_values=FLAGS,
-        )
-        orig_flags = FLAGS.flag_values_dict()
-        launcher.job_cls.define_flags(FLAGS)
-        output_lines = []
-        FLAGS._render_flag_list(
-            [FLAGS[name] for name in FLAGS if name not in orig_flags], output_lines
-        )
+    action = parse_action(sys.argv, options=_ACTIONS, default="start")
+    cfg = _get_launcher_or_exit(action=action, require_runner=False, flag_values=fv)
+    orig_flags = fv.flag_values_dict()
+    define_flags(cfg, fv)
+
+    # If we are able to infer runner (e.g., user specified `--instance_type`), also amend the
+    # helpstring with more flags (which would be printed if user also specified `--help`).
+    if fv.runner_name:
+        output_lines = [f"\nRunner flags [{fv.runner_name}]:\n"]
+        fv._render_flag_list([fv[name] for name in fv if name not in orig_flags], output_lines)
         launch_help = "\n".join(output_lines)
+    else:
+        runner_list = list(_get_runner_or_exit(fv).keys())
+        launch_help = (
+            f"\nPossible launchers are:\n\n{_registry_help_string(_LAUNCHERS)}\n"
+            f"Possible runners are:\n\n{runner_list}\n\n"
+            f"Specify --runner_module to use different runners.\n"
+            f"Specify --runner_name to see additional flags."
+        )
 
     app.usage = functools.partial(_wrapped_usage, usage=app.usage, launch_help=launch_help)
 
@@ -659,19 +739,14 @@ def _private_flags():
 def _wrapped_usage(
     *,
     usage: Callable,
-    launch_help: Optional[str],
+    launch_help: str,
     writeto_stdout: bool = False,
     **kwargs,
 ):
     """Wraps original usage by printing additional launch-specific help."""
     usage(writeto_stdout=writeto_stdout, **kwargs)
     of = sys.stdout if writeto_stdout else sys.stderr
-    if launch_help:
-        of.write(
-            f"\nFlags for the provided --instance_type={FLAGS.instance_type}:\n{launch_help}\n"
-        )
-    else:
-        of.write("\nPass --instance_type to see additional flags.\n")
+    of.write(launch_help)
 
 
 if __name__ == "__main__":