+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Changes
   * Upgrade Jax from 0.4.37 to 0.4.38
   * Removes all QRM (queued resource manager) codepaths from `axlearn.cloud.gcp`.
+  * Introduces `named_runner_configs`. See `axlearn gcp launch --help` for details.
 
 ## 0.1.5